+++ doc/motivation.md
@@ -1,0 +1,44 @@
+# 为什么做 MCS（MyClashShell）
+
+本文说明我做这套工具的动机和几处刻意的设计取舍。它与 [README](../README.md) 分工不同：README 侧重「是什么、怎么用」，这里侧重「为什么这样设计」。
+
+## Clash Core 很好，但只靠 REST 不够顺手
+
+Clash Core 作为代理内核非常成熟，控制面也主要通过 **REST API** 暴露出来。这很灵活，但日常若每次都自己记地址、拼路径、用 `curl` 或零散文案去调接口，成本高、也容易忘。
+
+MCS 把常见操作收拢成 **`myclash` 的统一入口**，并在需要时用终端里的 **TUI**（`python -m tui`）连本地 API，把选节点、看状态这类事变成可重复、可形成肌肉记忆的操作，而不是每次都从零写一遍 HTTP 调用。
+
+## 代理开关要分得清：终端与图形环境
+
+我希望「开代理」不是模糊的全局状态，而是能按场景控制 **哪些环境里的程序会走代理**：
+
+- **`myclash shell on` / `off`**：作用于当前终端会话的 `http_proxy` / `https_proxy`，适合命令行工具、构建与脚本。
+- **`myclash window on` / `off`**：通过 GNOME 的系统代理，让图形程序走 Clash 暴露的 HTTP 代理端口。
+
+这样可以只开终端、只开桌面，或按需组合，而不必猜测某个 GUI 是否读了环境变量、某个终端是否继承了代理。
+
+## 为什么不用 TUN 模式：我更在意可控性
+
+我没有把 **TUN 模式**（在系统层统一接管流量）作为默认或主推路径，这是有意为之。
+
+TUN 的优势是「省事」：很多程序不用单独配置就能被统一下发。但代价是流量边界变模糊——**谁在走代理、为什么走、出问题从哪里查**，往往更费神。我更希望：
+
+- **边界清晰**：走代理的程序要么遵守环境变量/系统代理，要么由我显式配置；而不是在内核/虚拟网卡层一锅端。
+- **行为可预期**：减少与本地服务、局域网或个别应用路由策略的意外耦合；排障时更容易判断是代理链还是应用本身。
+- **部署与权限**：TUN 通常涉及更高权限和更多系统侧配置；我更倾向以 Clash 的 HTTP/SOCKS 端口配合显式开关为主，把复杂度放在我看得见的地方。
+
+我接受因此带来的取舍：不像 TUN 那样「一键全局」，需要为部分场景单独配代理或开关。对我来说，**可预测、可切换** 比 **一键全覆盖** 更重要。
+
+## 尽量不依赖图形界面来使用
+
+图形化的 Dashboard 或客户端在不少场景下很好用，但把它们当作**唯一**入口会带来两类问题：
+
+一是 **环境差异**：不同发行版、版本和桌面环境下，图形栈的行为和稳定性参差不齐；远程登录、自动化脚本里也不如终端一致。
+
+二是 **无头设备**：服务器、嵌入式或我主力用 SSH 的机器往往 **没有图形界面**。若使用流程必须先开浏览器或桌面应用，这条路在这些设备上就等于走不通。
+
+因此 MCS 的主路径是 **Shell 命令 + 可选 TUI**；`window on` 影响的是桌面环境的系统代理设置，那是「让 GUI 程序受益」的一种开关，而不是「必须先打开某个图形面板才能开始用」的前置条件。
+
+---
+
+若你也有类似场景，欢迎对照 [Quickstart](./quick_start.md) 试用；具体配置细节见 [clash_config.md](./clash_config.md)。
