+++ tui/__init__.py
@@ -1,0 +1,5 @@
+"""MyClashShell terminal UI (Textual + Clash / mihomo REST API)."""
+
+from tui.app import ClashTuiApp, main
+
+__all__ = ["ClashTuiApp", "main"]
