+++ tui/app.py
@@ -1,9 +1,7 @@
-#!/usr/bin/env python3
-"""MyClashShell terminal UI (Textual + Clash / mihomo REST API)."""
+"""MyClashShell Textual application (main screen)."""
 
 from __future__ import annotations
 
-import concurrent.futures
 import os
 import signal
 import sys
@@ -11,385 +9,38 @@ import threading
 import time
 from collections import deque
 from typing import Any, Callable
-from urllib.parse import quote
 
 import requests
+from textual import on
+from textual.app import App, ComposeResult
+from textual.binding import Binding
+from textual.events import Key
+from textual.containers import Horizontal, ItemGrid, Vertical, VerticalScroll
+from textual.widgets import (
+    Button,
+    ContentSwitcher,
+    Footer,
+    Input,
+    Label,
+    ListItem,
+    ListView,
+    RichLog,
+    Select,
+    Static,
+    Switch,
+)
 
-try:
-    from textual import on
-    from textual.app import App, ComposeResult
-    from textual.binding import Binding
-    from textual.events import Key
-    from textual.containers import Horizontal, ItemGrid, Vertical, VerticalScroll
-    from textual.widgets import (
-        Button,
-        ContentSwitcher,
-        DataTable,
-        Footer,
-        Input,
-        Label,
-        ListItem,
-        ListView,
-        RichLog,
-        Select,
-        Static,
-        Switch,
-    )
-except ImportError:
-    print(
-        "Textual is required. Install with:\n"
-        "  ${MYCLASH_ROOT_PWD}/venv/bin/pip install textual",
-        file=sys.stderr,
-    )
-    sys.exit(1)
-
-
-VIEW_IDS = [
-    "view-overview",
-    "view-proxies",
-    "view-rules",
-    "view-connections",
-    "view-config",
-    "view-logs",
-]
-
-
-def _fmt_bytes(n: float) -> str:
-    for unit in ("B", "KB", "MB", "GB", "TB"):
-        if abs(n) < 1024.0:
-            return f"{n:.1f} {unit}" if unit != "B" else f"{int(n)} {unit}"
-        n /= 1024.0
-    return f"{n:.1f} PB"
-
-
-def _fmt_rate(bps: float) -> str:
-    return _fmt_bytes(bps) + "/s"
-
-
-def _truncate(text: str, max_len: int) -> str:
-    if max_len <= 0:
-        return ""
-    if len(text) <= max_len:
-        return text
-    if max_len <= 3:
-        return text[:max_len]
-    return text[: max_len - 3] + "..."
-
-
-def _sparkline(values: list[float], width: int = 48) -> str:
-    if not values:
-        return "—"
-    blocks = "▁▂▃▄▅▆▇█"
-    chunk = max(1, len(values) // width)
-    samples = [sum(values[i : i + chunk]) / chunk for i in range(0, len(values), chunk)][-width:]
-    if not samples:
-        return "—"
-    lo, hi = min(samples), max(samples)
-    if hi <= lo:
-        return blocks[4] * len(samples)
-    return "".join(blocks[int((v - lo) / (hi - lo) * 7.999)] for v in samples)
-
-
-def _cfg_int(value: Any) -> int:
-    if value is None:
-        return 0
-    if isinstance(value, bool):
-        return int(value)
-    try:
-        return int(float(value))
-    except (TypeError, ValueError):
-        return 0
-
-
-def normalize_runtime_config(raw: Any) -> dict[str, Any]:
-    """统一不同内核/版本的 GET /configs 结构（扁平或包在 config 里）。"""
-    if not isinstance(raw, dict):
-        return {}
-    data = raw
-    for wrap in ("config", "Config", "data", "Data"):
-        inner = data.get(wrap)
-        if isinstance(inner, dict) and any(
-            k in inner for k in ("port", "mixed-port", "mode", "log-level", "allow-lan")
-        ):
-            data = dict(inner)
-            break
-    return data
-
-
-class ClashClient:
-    def __init__(self):
-        self.base_url = os.getenv("MYCLASH_API", "http://127.0.0.1:9090").rstrip("/")
-        self.secret = os.getenv("MYCLASH_SECRET", "").strip()
-
-    def _headers(self) -> dict[str, str]:
-        headers = {"Content-Type": "application/json"}
-        if self.secret:
-            headers["Authorization"] = f"Bearer {self.secret}"
-        return headers
-
-    def _get_json(self, path: str, params: dict | None = None, timeout: float = 4) -> Any:
-        r = requests.get(
-            f"{self.base_url}{path}",
-            headers=self._headers(),
-            params=params or {},
-            timeout=timeout,
-        )
-        r.raise_for_status()
-        if not r.content:
-            return {}
-        return r.json()
-
-    def get_proxies(self) -> dict:
-        return self._get_json("/proxies").get("proxies", {})
-
-    def get_connections(self) -> dict[str, Any]:
-        return self._get_json("/connections")
-
-    def get_rules(self) -> dict[str, Any]:
-        return self._get_json("/rules")
-
-    def get_configs(self) -> dict[str, Any]:
-        raw = self._get_json("/configs")
-        return normalize_runtime_config(raw)
-
-    def patch_configs(self, payload: dict[str, Any]) -> None:
-        r = requests.patch(
-            f"{self.base_url}/configs",
-            headers=self._headers(),
-            json=payload,
-            timeout=10,
-        )
-        if r.status_code >= 400:
-            detail = (r.text or "").strip() or r.reason or str(r.status_code)
-            raise RuntimeError(detail)
-        r.raise_for_status()
-
-    def test_delay(self, proxy_name: str, test_url: str, timeout_ms: int):
-        encoded_name = quote(proxy_name, safe="")
-        response = requests.get(
-            f"{self.base_url}/proxies/{encoded_name}/delay",
-            params={"url": test_url, "timeout": timeout_ms},
-            headers=self._headers(),
-            timeout=max(4, timeout_ms / 1000 + 1),
-        )
-        response.raise_for_status()
-        return response.json().get("delay")
-
-    def select_proxy(self, group_name: str, proxy_name: str) -> None:
-        encoded_group = quote(group_name, safe="")
-        response = requests.put(
-            f"{self.base_url}/proxies/{encoded_group}",
-            json={"name": proxy_name},
-            headers=self._headers(),
-            timeout=4,
-        )
-        response.raise_for_status()
-
-
-class TuiState:
-    def __init__(self, client: ClashClient, preferred_group: str | None = None):
-        self.client = client
-        self.groups: list[str] = []
-        self.group_idx = 0
-        self.nodes: list[str] = []
-        self.current_node = ""
-        self.selected_idx = 0
-        self.delays: dict[str, int | None] = {}
-        self.testing = False
-        self.last_error = ""
-        self.filter_text = ""
-        self.test_url = os.getenv(
-            "MYCLASH_TUI_TEST_URL", "https://www.gstatic.com/generate_204"
-        )
-        self.test_timeout_ms = int(os.getenv("MYCLASH_TUI_TIMEOUT_MS", "2500"))
-        self._lock = threading.Lock()
-        self._abort_delay_test = threading.Event()
-        self._preferred_group = preferred_group or os.getenv("MYCLASH_TUI_GROUP", "").strip()
-
-    def _pick_initial_group(self) -> int:
-        if not self.groups:
-            return 0
-        if self._preferred_group and self._preferred_group in self.groups:
-            return self.groups.index(self._preferred_group)
-        if "GLOBAL" in self.groups:
-            return self.groups.index("GLOBAL")
-        return 0
-
-    def display_nodes(self) -> list[str]:
-        if not self.filter_text.strip():
-            return list(self.nodes)
-        needle = self.filter_text.strip().lower()
-        return [n for n in self.nodes if needle in n.lower()]
-
-    @staticmethod
-    def _index_in_proxy_list(name: str, names: list[str]) -> int | None:
-        """在 all 列表里定位节点下标（与 API 的 now 对齐，忽略大小写差异）。"""
-        n = (name or "").strip()
-        if not n:
-            return None
-        if n in names:
-            return names.index(n)
-        low = n.lower()
-        for i, item in enumerate(names):
-            if item.lower() == low:
-                return i
-        return None
-
-    def sync_selection_to_api_current(self) -> None:
-        """键盘选中下标与 GET /proxies 里当前组的 now 一致（过滤后若在列表中也会对准）。"""
-        visible = self.display_nodes()
-        if not visible:
-            self.selected_idx = 0
-            return
-        idx = self._index_in_proxy_list(self.current_node, visible)
-        if idx is not None:
-            self.selected_idx = idx
-            return
-        if self.selected_idx >= len(visible):
-            self.selected_idx = max(0, len(visible) - 1)
-        elif self.selected_idx < 0:
-            self.selected_idx = 0
-
-    def refresh_groups_and_nodes(self) -> None:
-        proxies = self.client.get_proxies()
-        selector_types = {"Selector", "URLTest", "Fallback", "LoadBalance"}
-        groups: list[str] = []
-        for name, data in proxies.items():
-            if data.get("type") in selector_types and isinstance(data.get("all"), list) and data["all"]:
-                groups.append(name)
-        groups.sort()
-        if not groups:
-            raise RuntimeError("No proxy selector groups found from Clash REST API.")
-
-        previous_group = self.groups[self.group_idx] if self.groups else None
-        self.groups = groups
-        if previous_group in self.groups:
-            self.group_idx = self.groups.index(previous_group)
-        else:
-            self.group_idx = self._pick_initial_group()
-
-        group_name = self.groups[self.group_idx]
-        group_data = proxies.get(group_name, {})
-        self.nodes = list(group_data.get("all", []))
-        raw_now = group_data.get("now")
-        if raw_now is None:
-            raw_now = group_data.get("Now")
-        self.current_node = str(raw_now).strip() if raw_now is not None else ""
-        self.sync_selection_to_api_current()
-
-    def cycle_group(self, step: int) -> None:
-        if not self.groups:
-            return
-        self.group_idx = (self.group_idx + step) % len(self.groups)
-        self.delays = {}
-        self.refresh_groups_and_nodes()
-
-    def start_delay_test(self, on_done: Callable[[], None] | None = None) -> None:
-        if self.testing or not self.nodes:
-            return
-        self._abort_delay_test.clear()
-        self.testing = True
-        self.last_error = ""
-        nodes = list(self.nodes)
-
-        def worker():
-            result: dict[str, int | None] = {}
-            try:
-                executor = concurrent.futures.ThreadPoolExecutor(max_workers=8)
-                try:
-                    future_map = {
-                        executor.submit(
-                            self.client.test_delay, node, self.test_url, self.test_timeout_ms
-                        ): node
-                        for node in nodes
-                    }
-                    pending = set(future_map.keys())
-                    while pending:
-                        if self._abort_delay_test.is_set():
-                            break
-                        done, pending = concurrent.futures.wait(
-                            pending,
-                            timeout=0.4,
-                            return_when=concurrent.futures.FIRST_COMPLETED,
-                        )
-                        for fut in done:
-                            node = future_map[fut]
-                            try:
-                                result[node] = fut.result()
-                            except Exception:
-                                result[node] = None
-                finally:
-                    executor.shutdown(wait=False, cancel_futures=True)
-            except Exception as exc:
-                self.last_error = str(exc)
-            with self._lock:
-                self.delays = result
-                self.testing = False
-            if on_done:
-                on_done()
-
-        threading.Thread(target=worker, daemon=True).start()
-
-    def select_current(self) -> None:
-        visible = self.display_nodes()
-        if not visible:
-            return
-        group_name = self.groups[self.group_idx]
-        node = visible[self.selected_idx]
-        self.client.select_proxy(group_name, node)
-        self.current_node = node
-
-
-def _delay_text(delay: int | None) -> str:
-    if delay is None:
-        return "-- ms"
-    return f"{delay} ms"
-
-
-def _delay_style(delay: int | None) -> str:
-    if delay is None:
-        return "dim"
-    if delay <= 250:
-        return "green"
-    if delay <= 600:
-        return "yellow"
-    return "red"
-
-
-class ProxyCard(Static):
-    def __init__(self, node_name: str, **kwargs):
-        kwargs.setdefault("markup", True)
-        super().__init__(**kwargs)
-        self.node_name = node_name
-
-    DEFAULT_CSS = """
-    ProxyCard {
-        height: auto;
-        min-height: 5;
-        padding: 0 1;
-        border: round $primary;
-        background: $surface;
-    }
-    ProxyCard.selected {
-        border: heavy $accent;
-        background: $primary 18%;
-    }
-    """
-
-    def set_content(self, selected: bool, current: bool, delay: int | None) -> None:
-        prefix = "* " if current else "  "
-        name_line = _truncate(prefix + self.node_name, 28)
-        delay_line = _delay_text(delay)
-        dstyle = _delay_style(delay)
-        self.border_title = name_line
-        if selected:
-            self.add_class("selected")
-        else:
-            self.remove_class("selected")
-        self.update(f"[{dstyle}]{delay_line}[/]")
+from tui.client import ClashClient
+from tui.config_api import _cfg_int
+from tui.constants import VIEW_IDS
+from tui.formatting import _fmt_bytes, _fmt_rate, _sparkline, _truncate
+from tui.state import TuiState
+from tui.widgets import ProxyNodeButton
 
 
 class ClashTuiApp(App[None]):
+    theme = "nord"
+
     CSS = """
     Screen {
         background: $surface;
@@ -414,16 +65,27 @@ class ClashTuiApp(App[None]):
         border: none;
         padding: 0;
     }
-    .sidebar-brand {
+    #sidebar-nav ListItem {
+        padding: 1 1;
+        margin: 0 1;
+        height: auto;
+        min-height: 3;
+    }
+    #sidebar-nav Label {
         text-style: bold;
-        color: $accent;
-        padding: 0 1 1 1;
+    }
+    #sidebar-nav ListItem:hover {
+        background: $foreground 8%;
+    }
+    #sidebar-nav ListItem.-highlight {
+        background: $primary 28%;
+        text-style: bold;
     }
     #main {
         width: 1fr;
         height: 100%;
         layout: vertical;
-        padding: 0 1;
+        padding: 0 0 0 1;
     }
     #main-views {
         height: 1fr;
@@ -437,21 +99,21 @@ class ClashTuiApp(App[None]):
     .page-header {
         height: auto;
         layout: horizontal;
-        padding: 1 0;
+        padding: 0 0 1 0;
         align-vertical: middle;
     }
     .page-title {
         width: 1fr;
         text-style: bold;
-        color: $text;
-    }
-    #search {
-        width: 36;
+        color: $accent;
+        margin-bottom: 0;
+        padding-bottom: 1;
+        border-bottom: wide $boost;
     }
     #group-line {
         height: auto;
         color: $text-muted;
-        margin-bottom: 1;
+        margin-bottom: 0;
     }
     #proxy-scroll {
         height: 1fr;
@@ -459,24 +121,24 @@ class ClashTuiApp(App[None]):
     }
     #proxy-grid {
         height: auto;
-        grid-gutter: 1 1;
+        grid-gutter: 0 1;
     }
     #status-line {
         height: auto;
         color: $text-muted;
-        padding-top: 1;
+        padding-top: 0;
     }
     .stat-row {
         layout: horizontal;
         height: auto;
-        margin: 1 0;
-        grid-gutter: 1;
+        margin: 0 0 1 0;
+        grid-gutter: 0;
     }
     .stat-card {
         width: 1fr;
         height: auto;
-        min-height: 5;
-        padding: 0 1;
+        min-height: 3;
+        padding: 0;
         border: round $primary;
         background: $panel;
     }
@@ -488,33 +150,65 @@ class ClashTuiApp(App[None]):
     }
     #overview-chart {
         height: auto;
-        margin-top: 1;
+        margin-top: 0;
         color: $text-muted;
     }
     #overview-chart:focus {
         background-tint: $foreground 8%;
     }
-    #rules-table, #conn-table {
+    .config-sheet {
         height: 1fr;
-        border: tall $primary;
+        layout: vertical;
+        border: round $boost;
+        padding: 0 1 1 1;
+        margin-top: 0;
+        background: $panel;
+    }
+    .config-section-title {
+        height: auto;
+        color: $text-muted;
+        text-style: bold;
+        margin-top: 1;
+        margin-bottom: 0;
+    }
+    .config-section-title:first-of-type {
+        margin-top: 0;
+    }
+    .config-section {
+        height: auto;
+        layout: vertical;
     }
     .cfg-row {
         height: auto;
-        margin-bottom: 1;
+        margin-bottom: 0;
         layout: horizontal;
         align-vertical: middle;
     }
     .cfg-label {
-        width: 18;
+        width: 14;
         color: $text-muted;
     }
     #cfg-api-status {
         height: auto;
         color: $text-muted;
-        margin-bottom: 1;
+        text-style: italic;
+        margin-bottom: 0;
+    }
+    #cfg-bind {
+        color: $text-muted;
+        text-style: italic;
+    }
+    Input:focus {
+        border: tall $accent;
+    }
+    Select:focus {
+        border: tall $accent;
+    }
+    Switch:focus {
+        border: tall $accent;
     }
     #cfg-apply {
-        margin-top: 1;
+        margin-top: 0;
         width: auto;
     }
     #log-panel {
@@ -546,6 +240,8 @@ class ClashTuiApp(App[None]):
 
     def __init__(self, preferred_group: str | None = None):
         super().__init__()
+        _theme = os.getenv("MYCLASH_TUI_THEME", "").strip() or "nord"
+        self.theme = _theme
         self._preferred_group = preferred_group
         self._client = ClashClient()
         self._state = TuiState(self._client, preferred_group=preferred_group)
@@ -555,24 +251,18 @@ class ClashTuiApp(App[None]):
         self._prev_up = 0
         self._down_hist: deque[float] = deque(maxlen=120)
         self._up_hist: deque[float] = deque(maxlen=120)
-        self._conn_filter = ""
         self._log_filter = ""
         self._log_stop = threading.Event()
         self._log_started = False
         self._overview_err = ""
-        self._conn_err = ""
-        self._rules_err = ""
         self._last_runtime_config: dict[str, Any] | None = None
 
     def compose(self) -> ComposeResult:
         with Horizontal(id="root"):
             with Vertical(id="sidebar"):
-                yield Static("MCS", classes="sidebar-brand")
                 yield ListView(
                     ListItem(Label("概览")),
                     ListItem(Label("代理")),
-                    ListItem(Label("规则")),
-                    ListItem(Label("连接")),
                     ListItem(Label("配置")),
                     ListItem(Label("日志")),
                     id="sidebar-nav",
@@ -601,75 +291,69 @@ class ClashTuiApp(App[None]):
                                 yield Static("—", id="ov-nb", classes="stat-value")
                         yield Static("", id="overview-chart", markup=False)
                     with Vertical(id="view-proxies", classes="view-pane"):
-                        with Horizontal(classes="page-header"):
-                            yield Static("代理", classes="page-title")
-                            yield Input(placeholder="搜索节点…", id="search")
+                        yield Static("代理", classes="page-title")
                         yield Static(id="group-line", markup=False)
                         with VerticalScroll(id="proxy-scroll", can_focus=False):
                             yield ItemGrid(
                                 id="proxy-grid",
-                                min_column_width=26,
+                                min_column_width=12,
                                 regular=True,
                             )
-                    with Vertical(id="view-rules", classes="view-pane"):
-                        with Horizontal(classes="page-header"):
-                            yield Static("规则", classes="page-title")
-                        yield DataTable(id="rules-table", zebra_stripes=True)
-                    with Vertical(id="view-connections", classes="view-pane"):
-                        with Horizontal(classes="page-header"):
-                            yield Static("连接", classes="page-title")
-                            yield Input(placeholder="过滤…", id="conn-filter")
-                        yield DataTable(id="conn-table", zebra_stripes=True)
                     with Vertical(id="view-config", classes="view-pane"):
                         with Horizontal(classes="page-header"):
                             yield Static("配置", classes="page-title")
-                        yield Static("", id="cfg-api-status", markup=False)
-                        with VerticalScroll():
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("HTTP 端口", classes="cfg-label")
-                                yield Input("0", id="cfg-port", placeholder="port")
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("SOCKS5", classes="cfg-label")
-                                yield Input("0", id="cfg-socks", placeholder="socks-port")
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("Mixed", classes="cfg-label")
-                                yield Input("7890", id="cfg-mixed", placeholder="mixed-port")
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("Redir", classes="cfg-label")
-                                yield Input("0", id="cfg-redir", placeholder="redir-port")
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("TProxy", classes="cfg-label")
-                                yield Input("0", id="cfg-tproxy", placeholder="tproxy-port")
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("模式", classes="cfg-label")
-                                yield Select(
-                                    [("Rule", "rule"), ("Global", "global"), ("Direct", "direct")],
-                                    id="cfg-mode",
-                                    allow_blank=False,
-                                )
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("日志级别", classes="cfg-label")
-                                yield Select(
-                                    [
-                                        ("Debug", "debug"),
-                                        ("Info", "info"),
-                                        ("Warning", "warning"),
-                                        ("Error", "error"),
-                                        ("Silent", "silent"),
-                                    ],
-                                    id="cfg-loglevel",
-                                    allow_blank=False,
-                                )
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("允许局域网", classes="cfg-label")
-                                yield Switch(value=False, id="cfg-lan")
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("IPv6", classes="cfg-label")
-                                yield Switch(value=False, id="cfg-ipv6")
-                            with Horizontal(classes="cfg-row"):
-                                yield Label("绑定地址", classes="cfg-label")
-                                yield Static("—", id="cfg-bind", markup=False)
-                            yield Button("应用到内核 (PATCH /configs)", id="cfg-apply", variant="primary")
+                        with Vertical(classes="config-sheet"):
+                            yield Static("", id="cfg-api-status", markup=False)
+                            with VerticalScroll():
+                                yield Static("端口", classes="config-section-title")
+                                with Vertical(classes="config-section"):
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("HTTP 端口", classes="cfg-label")
+                                        yield Input("0", id="cfg-port", placeholder="port")
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("SOCKS5", classes="cfg-label")
+                                        yield Input("0", id="cfg-socks", placeholder="socks-port")
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("Mixed", classes="cfg-label")
+                                        yield Input("7890", id="cfg-mixed", placeholder="mixed-port")
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("Redir", classes="cfg-label")
+                                        yield Input("0", id="cfg-redir", placeholder="redir-port")
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("TProxy", classes="cfg-label")
+                                        yield Input("0", id="cfg-tproxy", placeholder="tproxy-port")
+                                yield Static("运行", classes="config-section-title")
+                                with Vertical(classes="config-section"):
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("模式", classes="cfg-label")
+                                        yield Select(
+                                            [("Rule", "rule"), ("Global", "global"), ("Direct", "direct")],
+                                            id="cfg-mode",
+                                            allow_blank=False,
+                                        )
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("日志级别", classes="cfg-label")
+                                        yield Select(
+                                            [
+                                                ("Debug", "debug"),
+                                                ("Info", "info"),
+                                                ("Warning", "warning"),
+                                                ("Error", "error"),
+                                                ("Silent", "silent"),
+                                            ],
+                                            id="cfg-loglevel",
+                                            allow_blank=False,
+                                        )
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("允许局域网", classes="cfg-label")
+                                        yield Switch(value=False, id="cfg-lan")
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("IPv6", classes="cfg-label")
+                                        yield Switch(value=False, id="cfg-ipv6")
+                                    with Horizontal(classes="cfg-row"):
+                                        yield Label("绑定地址", classes="cfg-label")
+                                        yield Static("—", id="cfg-bind", markup=False)
+                                yield Button("应用到内核 (PATCH /configs)", id="cfg-apply", variant="primary")
                     with Vertical(id="view-logs", classes="view-pane"):
                         with Horizontal(classes="page-header"):
                             yield Static("日志", classes="page-title")
@@ -703,12 +387,9 @@ class ClashTuiApp(App[None]):
         self.query_one("#overview-chart", Static).can_focus = True
         self.query_one("#sidebar-nav", ListView).focus()
         self.set_interval(1.0, self._tick_overview)
-        self.set_interval(1.0, self._tick_connections)
-        self._state.start_delay_test(on_done=lambda: self._safe_call_from_thread(self._after_delay_test))
+        self._start_delay_test_with_ui()
         if self._auto_refresh_s > 0:
             self.set_interval(float(self._auto_refresh_s), self._on_auto_timer)
-        self._prepare_rules_table()
-        self._prepare_conn_table()
         nav = self.query_one("#sidebar-nav", ListView)
         self._apply_sidebar_index(nav.index if nav.index is not None else 0)
 
@@ -719,18 +400,6 @@ class ClashTuiApp(App[None]):
     def action_help_quit(self) -> None:
         """Ctrl+C 字符路径（与 SIGINT 二选一或同时）：直接退出。"""
         self.exit()
-
-    def _prepare_rules_table(self) -> None:
-        table = self.query_one("#rules-table", DataTable)
-        if table.columns:
-            return
-        table.add_columns("#", "类型", "匹配", "策略")
-
-    def _prepare_conn_table(self) -> None:
-        table = self.query_one("#conn-table", DataTable)
-        if table.columns:
-            return
-        table.add_columns("主机", "进程", "下载", "上传", "↓/s", "↑/s", "链路", "规则")
 
     def _apply_sidebar_index(self, idx: int | None) -> None:
         if idx is None or not (0 <= idx < len(VIEW_IDS)):
@@ -754,11 +423,7 @@ class ClashTuiApp(App[None]):
             if vid == "view-overview":
                 self.query_one("#overview-chart", Static).focus()
             elif vid == "view-proxies":
-                self.query_one("#search", Input).focus()
-            elif vid == "view-rules":
-                self.query_one("#rules-table", DataTable).focus()
-            elif vid == "view-connections":
-                self.query_one("#conn-filter", Input).focus()
+                self.query_one("#proxy-grid", ItemGrid).focus()
             elif vid == "view-config":
                 self.query_one("#cfg-port", Input).focus()
             elif vid == "view-logs":
@@ -782,9 +447,7 @@ class ClashTuiApp(App[None]):
         self._focus_first_in_main()
 
     def _on_view_switched(self, vid: str) -> None:
-        if vid == "view-rules":
-            self.run_worker(self._load_rules_async(), group="rules", exit_on_error=False)
-        elif vid == "view-config":
+        if vid == "view-config":
             self.run_worker(self._load_config_async(), group="config", exit_on_error=False)
         elif vid == "view-logs":
             self._ensure_log_tail()
@@ -870,100 +533,6 @@ class ClashTuiApp(App[None]):
             f"上传 {_sparkline(list(self._up_hist))}"
         )
         self.query_one("#overview-chart", Static).update(chart)
-
-    def _tick_connections(self) -> None:
-        if self._current_view() != "view-connections":
-            return
-        self.run_worker(self._refresh_connections_async(), group="conn", exclusive=True, exit_on_error=False)
-
-    async def _refresh_connections_async(self) -> None:
-        try:
-            data = self._client.get_connections()
-            self._conn_err = ""
-        except Exception as exc:
-            self._conn_err = str(exc)
-            self._safe_call_from_thread(self._sync_main_footer)
-            return
-
-        conns = data.get("connections") or []
-        needle = self._conn_filter.strip().lower()
-        rows: list[tuple[str, ...]] = []
-        now_ms = time.time() * 1000
-        for c in conns:
-            meta = c.get("metadata") or {}
-            host = meta.get("host") or meta.get("destinationIP") or "—"
-            if meta.get("destinationPort"):
-                host = f"{host}:{meta.get('destinationPort')}"
-            proc = meta.get("processPath") or meta.get("process") or ""
-            if proc and "/" in proc:
-                proc = proc.rsplit("/", 1)[-1]
-            chains = " / ".join(c.get("chains") or []) or "—"
-            rule = str(c.get("rule") or c.get("rulePayload") or "—")
-            dl = int(c.get("download", 0))
-            ul = int(c.get("upload", 0))
-            start = c.get("start")
-            elapsed_s = 1.0
-            if isinstance(start, str):
-                try:
-                    elapsed_s = max(0.001, (now_ms - float(start)) / 1000.0)
-                except ValueError:
-                    pass
-            d_speed = dl / elapsed_s
-            u_speed = ul / elapsed_s
-            line = " ".join([host, proc, chains, rule]).lower()
-            if needle and needle not in line:
-                continue
-            rows.append(
-                (
-                    _truncate(str(host), 28),
-                    _truncate(str(proc), 12),
-                    _fmt_bytes(float(dl)),
-                    _fmt_bytes(float(ul)),
-                    _fmt_rate(d_speed),
-                    _fmt_rate(u_speed),
-                    _truncate(chains, 36),
-                    _truncate(rule, 24),
-                )
-            )
-
-        def apply():
-            table = self.query_one("#conn-table", DataTable)
-            table.clear()
-            for row in rows:
-                table.add_row(*row)
-
-        self._safe_call_from_thread(apply)
-        self._safe_call_from_thread(self._sync_main_footer)
-
-    async def _load_rules_async(self) -> None:
-        try:
-            payload = self._client.get_rules()
-            self._rules_err = ""
-        except Exception as exc:
-            self._rules_err = str(exc)
-            self._safe_call_from_thread(self._sync_main_footer)
-            return
-
-        rules = payload.get("rules")
-        if rules is None:
-            rules = payload if isinstance(payload, list) else []
-
-        def apply():
-            table = self.query_one("#rules-table", DataTable)
-            table.clear()
-            for i, r in enumerate(rules):
-                if isinstance(r, dict):
-                    table.add_row(
-                        str(i),
-                        _truncate(str(r.get("type", "")), 16),
-                        _truncate(str(r.get("payload", "")), 40),
-                        _truncate(str(r.get("proxy", "")), 20),
-                    )
-                else:
-                    table.add_row(str(i), "", str(r), "")
-
-        self._safe_call_from_thread(apply)
-        self._safe_call_from_thread(self._sync_main_footer)
 
     def _apply_config_form_from_dict(self, cfg: dict[str, Any], *, status_msg: str | None = None) -> None:
         """把 GET /configs 的扁平字段填回表单，并可选更新顶栏说明。"""
@@ -1070,10 +639,22 @@ class ClashTuiApp(App[None]):
     def _on_auto_timer(self) -> None:
         if self._state.testing:
             return
-        self._state.start_delay_test(on_done=lambda: self._safe_call_from_thread(self._after_delay_test))
+        self._start_delay_test_with_ui()
+
+    def _start_delay_test_with_ui(self) -> None:
+        self._state.start_delay_test(
+            on_done=lambda: self._safe_call_from_thread(self._after_delay_test),
+            on_progress=lambda: self._safe_call_from_thread(self._delay_test_progress),
+        )
+
+    def _delay_test_progress(self) -> None:
+        if self._current_view() == "view-proxies":
+            self._refresh_card_contents()
+            self._sync_proxy_chrome()
 
     def _after_delay_test(self) -> None:
         if self._current_view() == "view-proxies":
+            self._refresh_card_contents()
             self._sync_proxy_chrome()
         self._sync_main_footer()
 
@@ -1117,23 +698,19 @@ class ClashTuiApp(App[None]):
         err = self._state.last_error
         base = (
             "侧栏[↑↓]换页 [Enter]/[Tab]进主区  [ctrl+b]回侧栏  "
-            "[ctrl+i]搜索/过滤  [Esc]代理节点区  [ctrl+c]退出"
+            "[Esc]代理节点区  [ctrl+c]退出"
         )
         if vid == "view-proxies":
             extra = (
-                "  [↑↓←→/hjkl] 节点  [Enter] 切换  [r] 测速  [[]/]] 分组  [u] 同步"
+                "  [↑↓←→/hjkl] 节点  [Enter]/点击 切换  [r] 测速  [[]/]] 分组  [u] 同步"
                 + (f"  ·  {err}" if err else "")
             )
         elif vid == "view-overview":
             extra = f"  ·  {_truncate(self._overview_err, 80)}" if self._overview_err else ""
-        elif vid == "view-connections":
-            extra = f"  ·  {_truncate(self._conn_err, 80)}" if self._conn_err else ""
-        elif vid == "view-rules":
-            extra = f"  ·  {_truncate(self._rules_err, 80)}" if self._rules_err else ""
         elif vid == "view-config":
             extra = "  修改后点「应用」PATCH /configs"
         elif vid == "view-logs":
-            extra = "  日志流来自 GET /logs?level=info"
+            extra = "  [ctrl+i]过滤  日志流来自 GET /logs?level=info"
         else:
             extra = ""
         self.query_one("#status-line", Static).update(base + extra)
@@ -1152,20 +729,20 @@ class ClashTuiApp(App[None]):
         await grid.remove_children()
         visible = self._state.display_nodes()
         for i, node in enumerate(visible):
-            card = ProxyCard(node, id=f"pc-{i}")
-            await grid.mount(card)
+            btn = ProxyNodeButton(node, id=f"pb-{i}")
+            await grid.mount(btn)
         self._refresh_card_contents()
         self.call_after_refresh(self._scroll_to_selection)
 
     def _refresh_card_contents(self) -> None:
         visible = self._state.display_nodes()
-        cards = list(self.query("#proxy-grid ProxyCard"))
-        for i, card in enumerate(cards):
+        buttons = list(self.query("#proxy-grid ProxyNodeButton"))
+        for i, btn in enumerate(buttons):
             if i >= len(visible):
                 break
             name = visible[i]
             delay = self._state.delays.get(name)
-            card.set_content(
+            btn.set_node_state(
                 selected=(i == self._state.selected_idx),
                 current=(name == self._state.current_node),
                 delay=delay,
@@ -1176,10 +753,10 @@ class ClashTuiApp(App[None]):
         if not visible:
             return
         i = min(self._state.selected_idx, len(visible) - 1)
-        cards = list(self.query("#proxy-grid ProxyCard"))
-        if i < len(cards):
+        buttons = list(self.query("#proxy-grid ProxyNodeButton"))
+        if i < len(buttons):
             self.query_one("#proxy-scroll", VerticalScroll).scroll_to_widget(
-                cards[i], animate=False
+                buttons[i], animate=False
             )
 
     def _move(self, drow: int, dcol: int) -> None:
@@ -1229,7 +806,7 @@ class ClashTuiApp(App[None]):
     def action_retest(self) -> None:
         if not self._main_grid_commands_active():
             return
-        self._state.start_delay_test(on_done=lambda: self._safe_call_from_thread(self._after_delay_test))
+        self._start_delay_test_with_ui()
         self._sync_proxy_chrome()
         self._sync_main_footer()
 
@@ -1237,7 +814,7 @@ class ClashTuiApp(App[None]):
         if not self._main_grid_commands_active():
             return
         self._state.cycle_group(-1)
-        self._state.start_delay_test(on_done=lambda: self._safe_call_from_thread(self._after_delay_test))
+        self._start_delay_test_with_ui()
         await self._rebuild_cards_async()
         self._sync_proxy_chrome()
         self._sync_main_footer()
@@ -1246,7 +823,7 @@ class ClashTuiApp(App[None]):
         if not self._main_grid_commands_active():
             return
         self._state.cycle_group(1)
-        self._state.start_delay_test(on_done=lambda: self._safe_call_from_thread(self._after_delay_test))
+        self._start_delay_test_with_ui()
         await self._rebuild_cards_async()
         self._sync_proxy_chrome()
         self._sync_main_footer()
@@ -1259,18 +836,13 @@ class ClashTuiApp(App[None]):
             self._state.last_error = ""
         except Exception as exc:
             self._state.last_error = str(exc)
-        self._state.start_delay_test(on_done=lambda: self._safe_call_from_thread(self._after_delay_test))
+        self._start_delay_test_with_ui()
         await self._rebuild_cards_async()
         self._sync_proxy_chrome()
         self._sync_main_footer()
 
     def action_focus_search(self) -> None:
-        vid = self._current_view()
-        if vid == "view-proxies":
-            self.query_one("#search", Input).focus()
-        elif vid == "view-connections":
-            self.query_one("#conn-filter", Input).focus()
-        elif vid == "view-logs":
+        if self._current_view() == "view-logs":
             self.query_one("#logs-filter", Input).focus()
 
     def action_focus_sidebar(self) -> None:
@@ -1280,24 +852,35 @@ class ClashTuiApp(App[None]):
         if self._current_view() == "view-proxies":
             self.query_one("#proxy-grid", ItemGrid).focus()
 
-    @on(Input.Changed, "#search")
-    def filter_nodes(self, event: Input.Changed) -> None:
-        self._state.filter_text = event.value
-        self._state.sync_selection_to_api_current()
+    @on(Button.Pressed, ".proxy-node-btn")
+    def on_proxy_node_pressed(self, event: Button.Pressed) -> None:
+        if self._current_view() != "view-proxies":
+            return
+        wid = event.button.id or ""
+        if not wid.startswith("pb-"):
+            return
+        try:
+            idx = int(wid[3:])
+        except ValueError:
+            return
+        visible = self._state.display_nodes()
+        if not (0 <= idx < len(visible)):
+            return
+        self._state.selected_idx = idx
+        try:
+            self._state.select_current()
+        except Exception as exc:
+            self._state.last_error = str(exc)
+        self._refresh_card_contents()
         self._sync_proxy_chrome()
-        self._schedule_rebuild_cards()
-
-    @on(Input.Changed, "#conn-filter")
-    def filter_connections(self, event: Input.Changed) -> None:
-        self._conn_filter = event.value
-        self._tick_connections()
+        self._sync_main_footer()
 
     @on(Input.Changed, "#logs-filter")
     def filter_logs(self, event: Input.Changed) -> None:
         self._log_filter = event.value.strip()
 
 
-def main():
+def main() -> None:
     preferred_group = sys.argv[1] if len(sys.argv) > 1 else None
     app = ClashTuiApp(preferred_group=preferred_group)
 
@@ -1320,7 +903,3 @@ def main():
             signal.signal(signal.SIGINT, signal.SIG_DFL)
         except (OSError, ValueError):
             pass
-
-
-if __name__ == "__main__":
-    main()
