+++ tui/constants.py
@@ -1,0 +1,8 @@
+"""Sidebar / ContentSwitcher view ids."""
+
+VIEW_IDS = [
+    "view-overview",
+    "view-proxies",
+    "view-config",
+    "view-logs",
+]
