--- ubuntu/template/clash_dashboard.service
@@ -1,12 +1,0 @@
-[unit]
-Description=clash
-After=multi-user.target
-[Service]
-TimeoutStartSec=30
-ExecStart=/usr/bin/python3 ./ubuntu/scripts/dashboard.py
-WorkingDirectory=<|PLACEHOLD_1|>
-ExecStop=/bin/kill $MAINPID
-Restart=always
-RestartSec=10s
-[Install]
-WantedBy=multi-user.target
