+++ ubuntu/template/myclash.service
@@ -1,0 +1,17 @@
+[Unit]
+Description=MyClashShell (supervised Clash core)
+After=network-online.target
+Wants=network-online.target
+
+[Service]
+Type=simple
+TimeoutStartSec=30
+WorkingDirectory=<|PLACEHOLD_1|>
+Environment=MYCLASH_ROOT_PWD=<|PLACEHOLD_2|>
+ExecStart=<|PLACEHOLD_1|>/venv/bin/python3 <|PLACEHOLD_2|>/ubuntu/scripts/clash_supervisor.py
+ExecStop=/bin/kill -TERM $MAINPID
+Restart=always
+RestartSec=10s
+
+[Install]
+WantedBy=multi-user.target
